Extract RenderCardView and RenderTableView js_only
