+++ assets/src/domain/eventEditor/ui/datetimes/datesList/cardView/index.ts
@@ -1,1 +1,1 @@
-export { default as CardView } from './CardView';
+export { default as RenderCardView } from './RenderCardView';
