+++ assets/src/domain/eventEditor/ui/datetimes/datesList/tableView/index.ts
@@ -1,1 +1,1 @@
-export { default as TableView } from './TableView';
+export { default as RenderTableView } from './RenderTableView';
