+++ assets/src/domain/eventEditor/ui/tickets/ticketsList/tableView/RenderTableView.tsx
@@ -1,0 +1,15 @@
+import React, { Suspense } from 'react';
+
+import { LoadingNotice } from '@appDisplay/loadingNotice';
+
+const TableView = React.lazy(() => import(/* webpackChunkName: "tickets-table-view" */ './TableView'));
+
+const RenderTableView = () => {
+	return (
+		<Suspense fallback={<LoadingNotice />}>
+			<TableView />
+		</Suspense>
+	);
+};
+
+export default RenderTableView;
