+++ assets/src/domain/eventEditor/ui/tickets/ticketsList/tableView/index.ts
@@ -1,1 +1,1 @@
-export { default as TableView } from './TableView';
+export { default as RenderTableView } from './RenderTableView';
